move components.tsx down to app/
