+++ dp-game-webapp/app/game/results.tsx
@@ -8,7 +8,7 @@ import {
 
 import { Answer, AnsweredQuestion } from "./questions";
 import { ExponentialNumber } from "../exponentialNumber";
-import { GameContainer, PageContainer, PageTitle } from "./components";
+import { GameContainer, PageContainer, PageTitle } from "../components";
 
 const answerIcons: Record<Answer, ReactNode> = {
   [Answer.DecreaseSpend]: (
