+++ dp-game-webapp/app/game/start.tsx
@@ -11,7 +11,7 @@ import {
   PageContainer,
   PageDescription,
   PageTitle,
-} from "./components";
+} from "../components";
 
 export default function StartGame({
   impressions,
