+++ dp-game-webapp/app/game/validate.tsx
@@ -18,7 +18,7 @@ import {
   PageContainer,
   PageDescription,
   PageTitle,
-} from "./components";
+} from "../components";
 
 export default function Validate({
   conversionRate,
